Rename person test package to order
