+++ src/test/java/tracko/model/order/OrderContainsKeywordsPredicateTest.java
@@ -1,4 +1,4 @@
-package tracko.model.person;
+package tracko.model.order;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -9,8 +9,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import tracko.model.order.ItemQuantityPair;
-import tracko.model.order.OrderContainsKeywordsPredicate;
 import tracko.testutil.OrderBuilder;
 
 public class OrderContainsKeywordsPredicateTest {
